Move initial state update out of component
This ensures the initial state is correct from the start.

COSDK-1120
